Move JSON serialization to separate header to keep nlohmann_json private
ProFormaData.h no longer includes nlohmann/json.hpp directly. All JSON
serialization functions (to_json/from_json for ADL, toJSON/fromJSON
convenience functions) are moved to ProFormaDataJson.h.

This allows nlohmann_json to remain a PRIVATE dependency of OpenMS while
still providing JSON support for ProForma data structures to users who
need it.

Users who need JSON serialization should include ProFormaDataJson.h which
includes both ProFormaData.h and nlohmann/json.hpp.
